rename notes.ts to read.ts
